Download binaries in util.py and generate release.yml in build.sh
